Modify the crash bug.

+++ app/src/main/java/cn/edu/hit/application/ImageManager.java
@@ -94,8 +94,8 @@ public class ImageManager {
     public static Bitmap imageScale(Bitmap bitmap, float dst_w, float dst_h) {
         int src_w = bitmap.getWidth();
         int src_h = bitmap.getHeight();
-        float scale_w = ((float) dst_w) / src_w;
-        float scale_h = ((float) dst_h) / src_h;
+        float scale_w =  dst_w / src_w;
+        float scale_h =  dst_h / src_h;
         Matrix matrix = new Matrix();
         matrix.postScale(scale_w, scale_h);
         Bitmap dstbmp = Bitmap.createBitmap(bitmap, 0, 0, src_w, src_h, matrix,
